# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Photo Viewer API",
        "contact": {},
        "version": "dev"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes the session bound to the current access token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "session not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "user_login, user_email, user_role",
                        "schema": {
                            "$ref": "#/definitions/auth.UserInfoResponse"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "failed to get user info (incl. inactive user)",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Requires the refresh_token cookie. Revokes the current session and issues a new access token and cookie.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.AccessTokenResponse"
                        }
                    },
                    "401": {
                        "description": "missing authentication / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "already authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "failed to create tokens",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/auth/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify user email",
                "parameters": [
                    {
                        "description": "login (2-50), code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserVerifyCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "invalid metadata / validation",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "invalid code / already authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/photo/{photo_uuid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Photo info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo UUID",
                        "name": "photo_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access link secret (for protected photos)",
                        "name": "photo_access_secret",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.PhotoInfo"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "photo is not permitted",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "photo not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Delete a photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo UUID",
                        "name": "photo_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "invalid authorization (not the owner)",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Update photo metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo UUID",
                        "name": "photo_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JSON: title (≤50), description (≤200), created_at, took_at, access_level (private|protected|public)",
                        "name": "metadata",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/update.Response"
                        }
                    },
                    "400": {
                        "description": "invalid metadata / validation / invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "photo not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.AccessTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
        "auth.UserInfoResponse": {
            "type": "object",
            "properties": {
                "user_email": {
                    "type": "string"
                },
                "user_login": {
                    "type": "string"
                },
                "user_role": {
                    "type": "string"
                }
            }
        },
        "entity.AccessModifier": {
            "type": "string",
            "enum": [
                "private",
                "protected",
                "public"
            ],
            "x-enum-varnames": [
                "AccessModifierPrivate",
                "AccessModifierProtected",
                "AccessModifierPublic"
            ]
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
...
        "service.PhotoInfo": {
            "type": "object",
            "required": [
                "access_level"
            ],
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "access_level": {
                    "$ref": "#/definitions/entity.AccessModifier"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 200
                },
                "owner_login": {
                    "type": "string"
                },
                "photo_uuid": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.TagSmallInfo"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 50
                },
                "took_at": {
                    "type": "string"
                }
            }
        },
...
        "service.TagSmallInfo": {
            "type": "object",
            "properties": {
                "tag_name": {
                    "type": "string"
                },
                "tag_uuid": {
                    "type": "string"
                }
            }
        },
...
        "service.UserVerifyCredentials": {
            "type": "object",
            "required": [
                "login"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "login": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                }
            }
        },
...
        "update.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "photo_uuid": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer \u003caccess_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}